Adapt littleIMS to Cipango-3

+++ littleims-hss/src/main/java/org/cipango/ims/hss/db/hibernate/AbstractHibernateDao.java
@@ -74,7 +74,7 @@ public abstract class AbstractHibernateDao<T>
     
     public int count() 
     {
-    	return (Integer) criteria().setProjection(Projections.rowCount()).uniqueResult();
+    	return ((Long) criteria().setProjection(Projections.rowCount()).uniqueResult()).intValue();
 	}
     
     @SuppressWarnings("unchecked")
